\docType{package}
\name{tweet2r-package}
\alias{tweet2r-package}
\title{Access to Twitter Streaming API via R.} 
\description{
  A package base on \link{streamR} package which provides a serie of functions that allow R
  users to access Twitter's stream. It has been improved with two functions to capture tweets (setting up start and stop time) and store tweets in a postGIS data base.}

\author{
  Pau Aragó Galindo \email{parago@uji.es}
}
\seealso{
  \link{streamR}, \code{\link{tweet2r}},
  \code{\link{t2pgis}}
}
