\docType{package}
\name{tweet2r-package}
\alias{tweet2r-package}
\title{Access to Twitter Streaming API via R.} 
\description{
A package base on \link{streamR} package which provides a set of functions that allow R users to access Twitter's stream API. The package \code{\link{tweet2r}} has  a function to capture tweets using Twitter streaming API two functions to export the tweets to PostgreSQL , Sqlite shp, GML or KML and a function to get a summary of the tweets retrieved. Also there is a function to validate the JSON files where the tweets has been stored}

\author{
  Pau Aragó Galindo \email{parago@uji.es}
}
\seealso{
  \code{\link{streamR}}, \code{\link{tweet2r}},
  \code{\link{t2pgis}}

}
