\docType{package}
\name{tweet2r-package}
\alias{tweet2r-package}
\title{Access to Twitter Streaming API via R and perform a exploratory space-time analyis} 
\description{
A package base on \link{streamR} package which provides a set of functions that allow R users to access Twitter's stream API. The package \code{\link{tweet2r}} has  a function to capture tweets using Twitter streaming API two functions to export the tweets to PostgreSQL , Sqlite shp, GML or KML get Tweets as SpatialPointDataFrame, DataFrame,  a function to get a summary of the tweets retrieved and a exploratory space-time analyis. Also there is a function to validate the JSON files where the tweets has been stored}

\author{
  Pau Aragó Galindo \email{parago@uji.es}
}
\seealso{
  \code{\link{streamR}}, \code{\link{tweet2r}},
  \code{\link{t2pgis}}, \code{\link{t2sqlite}},
  \code{\link{t2gis}}, \code{\link{tsubset}},
  \code{\link{tglm}}, \code{\link{theatmap}},
  \code{\link{tspan}}, \code{\link{t2summary}},
  \code{\link{t2DataFrame}}, \code{\link{t2SpatialPointDataFrame}},

}
